🚀 Feature: Move playlist types to different file
