+++ client/src/features/playlist/playlistSlice.ts
@@ -1,41 +1,7 @@
 import { createSlice } from '@reduxjs/toolkit';
 import { getPlaylist } from './playlistThunks.ts';
 import toast from 'react-hot-toast';
-
-interface ApiStatus {
-  status: 'idle' | 'pending' | 'fulfilled' | 'rejected';
-  error: string;
-  statusCode?: number;
-}
-
-export interface Playlist {
-  id: string;
-  name: string;
-  img: {
-    id: string;
-    url: string;
-  };
-  user: {
-    id: string;
-    name: string;
-    img: {
-      id: string;
-      url: string;
-    };
-    role: 'user' | 'artist';
-  };
-  description?: string;
-  color: string;
-  length: number;
-  duration: number;
-}
-
-interface LibraryState {
-  data: Playlist | null;
-  api: {
-    getPlaylist: ApiStatus;
-  };
-}
+import { LibraryState } from './playlistTypes.ts';
 
 const initialState: LibraryState = {
   data: null,
